+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - fixed: (AVAX) Increased query lookback for transaction history to 5 minutes to fix missing transaction confirmations.
 - changed: Add redundancy to Cardano queries
 - changed: Increase `MAX_TRANSACTIONS` to 2500.
+- changed: (Hedera) Updated transaction query loop to account for mirror node limitations
 - fixed: (Hedera) Initial sync ratio inaccuracy
 
 ## 4.25.2 (2024-09-30)